update package declarations, move classes
