+++ edc-extensions/iatp/tx-iatp-sts-dim/src/main/resources/META-INF/services/org.eclipse.edc.spi.system.ServiceExtension
@@ -17,6 +17,6 @@
 #  SPDX-License-Identifier: Apache-2.0
 #################################################################################
 
-org.eclipse.tractusx.edc.iam.iatp.sts.dim.DimSecureTokenServiceExtension
-org.eclipse.tractusx.edc.iam.iatp.sts.dim.DimStsConfigurationExtension
+org.eclipse.tractusx.edc.iam.dcp.sts.RemoteTokenServiceClientExtension
+org.eclipse.tractusx.edc.iam.dcp.sts.StsClientConfigurationExtension
